Remove APIKeyForm component and KeyPrompt component, replacing their functionality with a new Settings component that manages API keys and user account information. Update Home and Settings routes to reflect these changes, ensuring a streamlined user experience for managing API keys and settings.
