+++ frontend/routes/Home.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import APIKeyManager from "@/frontend/components/APIKeyForm"
+import Settings from "@/frontend/components/Settings"
 import Chat from "@/frontend/components/Chat"
 import { v4 as uuidv4 } from "uuid"
 import { useAPIKeyStore } from "../stores/APIKeyStore"
@@ -49,7 +49,7 @@ export default function Home() {
   if (!hasRequiredKeys) {
     return (
       <div className="flex flex-col items-center justify-center w-full h-full max-w-3xl pt-10 pb-44 mx-auto">
-        <APIKeyManager />
+        <Settings />
       </div>
     )
   }
